refactor(MoveGenerator): MoveGenerator 클래스를 model 패키지로 이동
MoveGenerator 클래스를 model 패키지로 이동
