+++ src/main/java/model/MovesGenerator.java
@@ -1,4 +1,4 @@
-package utils;
+package model;
 
 import java.util.ArrayList;
 import java.util.List;
